self code review changes
